requirements.txt
Django==1.5.4
South==1.0.2
django_compressor==1.3
Pillow==2.4.0
sorl-thumbnail==11.12
django-crispy-forms==1.4.0
WebTest==2.0.9
django-webtest==1.7.5
djangorestframework==2.3.8
markdown==2.3.1 # Markdown support for the browsable API.
django-filter==0.7 # Filtering support
-e git://github.com/jmoiron/johnny-cache.git@johnny-cache-1.4.1#egg=johnny-cache-1.4.1
django-localflavor==1.0
celery==3.1.18
redis==2.9.1
django-celery==3.1.16
python-memcached==1.53
psycopg2==2.5.2
python-social-auth==0.2.6
django-tinymce==1.5.2
requests==2.3.0
feedly==0.10.10
django-haystack==2.3.1
Whoosh==2.6.0
factory_boy==2.4.1
djrill==1.3